Handle conflicting git rev in same repo
It turns out that a reasonable number of font sources share a repo, but
are based on different revisions.

This patch handles this by checking all of the referenced repos and revs
during source discovery, and marking conflicts. If a source is marked as
a conflict, it will use a rev-qualified directory when it is checked
out.